Ensuring cborHex output for signing keys also contain the verification key as per cardano-cli key convert-cardano-address-key outputs
